Moved NavBar to container file as it now is connnected to the store
