+++ src/containers/NavBar/style.scss
